feat(16h): hook useSavedViews + refactor SavedViewsBar; unificazione per clubs e opportunities
